requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
transformers
pdfplumber
python-dotenv
torch
accelerate
